Ui: Add primary ui skeleton.
The `MainUi` controls the state for the child Branding, MainViewPager, and
MainNavButtons child components.

When the user slides left or right in the MainViewPager the parent MainUi
state is updated to activate the corresponding NavButton. Conversly if the
user taps a NavButton the MainUi state is updated and MainViewPager is updated
with the correct View.

The whole view is styled with flexbox. If the buttons are too small change the
flex values in MainUi.