Redux: Save modal state in the store.
Ensure that the modal being rendered has the proper data.

When a `ListView` is rerendered with a modal open the props passed to child
components change to the first (or last?) list items props.

Now when a button is clicked to load a modal the data for that modal is passed
to the state. The modal then retrieves its props from the state. When that modal is closed the data is deleted.