Feature: Add login view with route to `MainUi`.
Use 'react-native-router-flux' to animate the press between `Login` and
`MainUi` views.

I suspect the transition is slow because `ViewPagerAndroid` must render all its views
before the transition is complete, and the maps view is slow to render.

Optimistically, the fix is to use routes in place of `ViewPagerAndroid`,
allowing for similar functionality but (much) better performance.