Feature: Use routes for `MainUi` navigation.
This removes `ViewPagerAndroid` leaving no platform specific code and paving
the way for iOS testing.

This is the first step in using routes for `MainUi`. A deeper refactor is
necessary for proper integration of react, redux, and
'react-native-router-flux' (rnrf).

Right now there are two bugs:

1. When `MapView` is navigated to multiple times the map has an image covering
part of the map. This does not happen the first time the view is rendered.

2. RNRF complains that "Key xxx is already defined" when changing scenes.