Feature: Add cross platform scrollable view pager.
Migrate from using routes to using 'react-native-scrollable-tab-view' for the
main ui buttons and views.

This is basically a cross platform `ViewPagerAndroid`.

To get the same look as before alteration of the render tab bar was necessary.

This is wired up to the view router and is compatible with the back button on
android.