Feature: add Facebook login.
Use the 'react-native-fbsdk' to handle logins and querying the graph api for
"contacts".

Use 'redux-saga' as middleware to connect the api calls into the `contacts`
reducer.

A huge caveat here is Facebook no longer allows querying for a list of all
friends from a user. In the Open Graph 2.0+ api `/me/friends` only returns the
friends that are using the same "app". The hack used here, querying
`invitable_friends`, is against the Facebook TOS.