Feature: Add loading actions for facebook sagas.
Use the react native `ActivityIndicator` to inform the user when a loading
action is in progress. Currently this applies to friends being loaded from
facebook within the contacts view.

Within a fetch saga before an async call is performed the state is changed to
indicate a load is in progress. This is passed to the view and presents a
loading indicator.