Api: Integrate stripe credit card payments.
A stripe payment happens in two steps:

1. The user entered credit card is sent to stripe for verification. If the
card is verified a "token" is sent within the response with the key of `id`.
If verification fails a stripe error is sent detailing the error.

2. The token is used to make a purchase with the verified credit card. The
purchase amount is specified without a decimal making 100 equal to 1 dollar.
Additionally a description is included in the purchase detailing how many
bevegrams were sent to whom.

Each step of this transaction has points of success and failure each of which
is passed to the state via the `put` commands.

The `PurchaseBeer` view adds the credit card information input with very
primitive verification that only checks the length of the entered characters.

The next step is integrating each saga step into the `PurchaseBeer` view.