Feature: Add ui for payment process.
When the "Confirm Purchase" button is pressed a new view is presented showing
the user each step is the purchase process. Each step is updated flowing from
the api request -> saga -> reducer -> state -> container -> component. If an
error occurs the relevant error message is shown to the user.

Credit card details are cleared when this modal is closed.