Storage: Load state from storage then render `InitialRouter`.
`redux-storage` does not set the initialState for the redux store but loads
the state async. This is not ideal. If a user is logged in they are shown the
login screen (the initialState is a logged out user) until the state is loaded
then transitioned to the contacts view.

This adds an app reducer with `isLoaded` set to false, when the state is
loaded the action `LOADING_COMPLETE` is dispatched. `isLoaded` is mapped to
`InitialRouter` which renders a loading screen (primitive for now) when
isLoaded is false and the login or contacts view depending on the users
saved login status.