Fix: Map render lags on android.
I explored many different ways to defer the rendering of a `MapView` until the
component is mounted but all were even worse than the original implementation.

This fix pre renders all views within the tab view, making the `MapView`
render before it is viewed. The trade off is more CPU and data usage for a
smoother experience on Android. iOS handles the map render much better, but
now it should be even smoother.