Feature: Add reducer for routes.
Very similar to the `modal` reducer.

This is setting up a transition from using modals to using routes for
secondary views that are entered into on pressing a button.

The core issue is a modal is too small on many screens. A route can take up
the full screen and is easy to exit with a swipe back, or pressing the back
button.