Feature: Wire up new routes functionality for `PurchaseBeer`.
Some differences between using a route vs using a modal.

If there is a `navBar` the height of the component must be compensated.

Components that may overflow must be wrapped in a scroll view.

Some styles were specific to a modal view and will require some tweaking.

Overall this works pretty smoothly.