Feature: Complete Adding and Saving Credit Cards.
In a nutshell this creates a customer with stripe and can add/remove/charge
each credit card.

How a user adds a credit card:

1. User inputs credit card details into the `AddCreditCard` view.

2. `AddCreditCard` performs simple checks on inputs

3. If the user does not have a stripe customer id one is generated.

3. Credit Card details are passed to stripe which verifies a credit card and
returns a token.

4. The token is added to the stripe customer and to the state.

5. The new card is rendered in the `PurchaseView`. The newest card is set as
the default (the card that will be charged).

How a user deletes a credit card:

1. Within the `Purchase` view user taps "Remove" next to the credit card and
waits until the card is removed from view.

How a user changes the Default card:

1. Within the `Purchase` view user taps the checkbox next to the card they
want to set as the default and waits until the "update" is complete.

How a user purchases a beer:

1. Within the `Purchase` view user taps "Confirm Purchase". The user is taken
to a view where they can see the steps in progress with a purchase. Once
complete they can return to the `Contacts` view via the "Close" button.