Test: Update stripe api and saga tests.
Currently the stripe saga tests only test the flow of the functions. It would
much better to test the state during the saga. This is difficult to setup
(time consuming to figure out how to mock the redux state) but this would be
the right way to test a saga.