Feature: Wire up back button for routes.
This forced a refactor of `GO_BACK_ROUTE` because `react-native-router-flux`
only allows a `backAndroid` function on the router and not individual scenes.
Previously, `GO_BACK_ROUTE` passed a payload of the previous route and the
next route. Now these values are stored in the state as `currentRoute` and
`previousRoute`.

Within the `goBackRoute` saga `routesThatDontGoBack` does what it says. The
current route is tested and if passed the Action goes back one route.

Additionally if the route is `MainUi` the `GOBACK_VIEW` action is dispatched
to navigate to a new tab.