Feature: Ensure "Purchase" button is always visible.
Move `bevLine` containing "Cancel" and "Purchase" buttons outside of
`RouteWithNavBarWrapper`.

Set the height of `bevButton` with `getButtonHeight` making buttons
with/without icons the same height. Also helps to determine the height of the
`bevLine` containing buttons.

Make the icon on the Purchase button the active card.

Add Font Awesome icons to `bevButton`.