Refactor: `Branding` is the new nav bar.
This is a total rewrite of the `BrandingComponent`.

The main goal here was to get the logo to be sized correctly. After much
struggling the best approach was to use fixed height and width values for the
logo png. I had avoided this because I thought flex sizing would work, but
flex sizing images is very buggy with react-native. Overall this adds some
complexity but offers much more predictable behavior.

Pro Tip:
Add `height` and `width` attributes to `Image`s if you don't want to spend
hours wondering why your images are behaving badly.