Feature: Move Purchasing/Sending in progress into route.
Take the existing implementation of `renderAttemptingPurchaseAndOrSend` and
move it into `PurchaseAndOrSendInProgress`.

The motivation for this change is to disallow a user to go back when
purchasing (or sending). This greatly reduces how many view states to keep
track, and ensures the user confirms the results of a purchase. Essentially
the only way a user can exit a Purchase is to press the "Close" button. This
makes managing the state easier.