Ui: Customize facebook login button.
This got weird. A custom button requires using `react-native-fbsdk`s
`LoginManager` which was no problem. But when a user is logged in a lot of
things happen causing the ui to stutter. The fix is to log a user in after the
facebook login is successful, but before the token is returned. With the token
the user and contacts can then be fetched. This makes the login "seem"
smoother.