Feature: Add user firebase integration.
Authenticate a firebase user using a facebook token. After a user logs in with
facebook they are authenticated as a firebase user using a facebook token.
Then the firebase credentials are saved to the state, which is then pushed to
the firebase database within the `users/${userFirebaseId}` table.

When a user logs out, they are logged out of firebase and facebook.