Feature: Add nosql database structure.
Add structure to how the database JSON is structured. Use interfaces to
standardize objects, functions to standardize url (a db location) location
strings. Basically make it hard to put unstructured data in the wrong
location.

Bevegrams are stored with a "UnixTime id" which is a concatenation of the unix
time and a unique id. A UnixTime id should allow for quick sorting and avoid
collisions.