Db: Add structure to urls.
The `Schema` object is what the firebase object looks like. Database locations
are matched against this schema before they are passed to external functions.
This imparts some structure on database transactions. It does not verify
types, but maybe it should?.

To get a url the schema location is passed as a string and verified, then if
an object is passed the object key is replaced with the object value. Ex:
`GetSchemaDbUrl("purchasedBevegrams.firebaseId.list", {firebaseId: "1234"})`
verifies that `purchasedBevegrams.firebase.list` is a valid location in `Schema` and returns
`purchasedBevegrams/1234/list` as the url.

On database nodes with potentially large sizes a summary of the data is
calculated on each transaction and stored in `.summary` and the objects stored
in `.list`. Ideally this means less sending and receiving of large json
objects for simple calculations, and quicker processing on the client.