Db: Add calls for reading/writing/updating db urls.
This adds the functions for reading and writing any/all data to firebase.
Essentially this uses the url getters from `db/schema` with firebase api
functions to execute db transactions.

Firebase is a nosql database (JSON) with a plethora of functions for db
transactions. To suit our needs the api is wrapped into three functions:

1. `PushToUrl`: Adds an object to a url with a firebase generated key that is
unique and time sortable. The id is returned with the intention of being used
within the state to store a copy of the list on disk. This is used on any node
with many objects, i.e. a users list of purchased bevegrams.

2. `UpdateNode`: Rewrites an object at a given url. A function is passed the
current value to use to create the updated object.

3. `ReadNode`: Returns an object at a given url.

This specifically does not use firebase events (`.on`), preferring more simple
transactions. More investigation is needed to determine if firebase events are
reliable within react native.

Along with adding bevegram objects to lists, summaries (metadata about a given
node) are generated and updated. Summary functions provide many uses:

1. They provide a snapshot of the data for more efficient calculations about
quantity and other relevant data. But this data is not guaranteed to be
correct.

2. They can provide comparison between the client and server without passing
potentially large json objects.

3. They provide a easily parseable format for data analysis.