State: Add sagas to pass data to `api/firebase`.
`sagas/firebase` combines data from `action.payload` and the state to create
the objects for writing to the database. Dates are "Stringified" using a
standard JSON format. Each of the functions returns a package with the id and
object that were written to the database to add these objects to the state.

`sagas/index` restructures the handling of dispatched actions. Some functions
depend on the results of other saga functions so they are combined here.