Refactor: Only allow purchasing and sending together.
Remove anything related to purchasing or sending separately. Don't store user
bevegrams in `UserState`.

This simplifies purchases from a user perspective because you can only send
and purchase together. It is not possible to keep your own bevegrams, they
immediately belong to the receiver.

Later we plan to add regifting so some logic was left as comments.