Ui: Add History tab
Render sorted history event keys into a list. Each list has different relevant
information so depending on what data the key is related to the format is
chosen accordingly.

Not creating an object with all lists combined is a deliberate choice. It
should more efficient to use a sorted list of keys and check each list to see
if it has a key than to combine all keys into a new object.