Feature: Add container for sorting contacts
At a high level this takes the contacts list and filters contacts according to
the selected sorting method, returning the sorted list of contacts for the
view to display.

For searching `fuse.js` seems to be the best option. It gives a fast fuzzy
search and is very configurable.

This also adds dispatches for updating the `contactsView` state.