Feature: Use firebase queue to send notifications from a server
Notifications via firebase cloud manager should be sent via a server because
putting the server key inside the app makes it very easy to steal, allowing an
attacker to send notifications on your behalf.

This is the first part of a larger refactor to put sensitive tasks onto a
server. This first attempt uses `firebase-queue` with an end goal of moving
all server side tasks to firebase's new cloud function feature.