Security: Don't package server side code into the app bundle
The idea here is to use the same codebase in two different applications. The
advantages are many (shared types, less reinventing the wheel, simplicity)
with the potential hazard of uploading server side code to an exposed app
bundle. Right now this script is the only line of defense.

This seems like a dirty hack, but this is what react-native itself uses to
ignore files for the bundle, so it seems like this will work for the
foreseeable future.