Feature: Listen and react to firebase db changes
Use firebase `child_changed` event to listen for the next change in the
database for a given url.

Listening is started within the `updateUserStateOnNextChange` saga. This saga
is passed actions to trigger on success and values to determine if the event
failed. On failure an error message is parsed from `user`, a message is passed
to the failure action and an optional alert shown.