Server: Move credit card sorting and identification to the backend.
To sort a customer's credit cards we add a `creationDate` metadata entry when
the card is created. When a customer is requested the `promiseCustomer` user
the `creationDate` to sort cards chronologically.

For security reasons we add a `generatedId` metadata entry when a card is
created using `node-uuid`. This id is passed to the client instead of the
actual stripe card id. This id is used to uniquely identify a card on the
client then passed to `convertGeneratedIdToStripeCardId` to get the actual
stripe card id for further processing. This is the last piece of stripe
security and now no stripe data is stored on the server.