Server: Add "Add Card to Customer" Queue actions
This is the first major part moving client functionality onto the server.

From a high level this takes the `addCreditCardToCustomer` saga and does the
same thing. The main difference is we don't have to check for an internet
connection after each action, we can basically assume that there will be no
connection errors.

This uses relatively new javascript `async` and `await` functionality for
cleaner code. This requires configuring babel to transpile for the current
node version.

`verifyUser` is a hack to ensure the sender is who the say they are.
`firebase-queue` does not provide a mechanism to identify the sender, so
before the queue is written to the client write a token to the
`verificationToken` node. The client can only write to their node, so if the
verification token sent does not match the verification token in the database
the client is not who they say they are and the process terminates.

`updateUser` is a helper function to ensure `user.lastModified` gets updated
to trigger the `child_changed` listener on the client. Failure to do this
results in a broken client state.