Feature: Move purchasing onto the backend
Differences from on client payment processing:

1. The status of the transaction must be passed back to the client. This is
performed in `updateStatus`. This also passes errors back to the client if the
purchase process fails for any reason.

2. Purchase packages are validated against the db purchase packages, failing
if a purchase package is not equal to one in the database.

3. Usage of async/await. Using sagas obfuscated the use of promises from
firebase functions. Now this is explicit with await everywhere. It would be
nice if firebase had types also to avoid these errors in the future.

4. The purchase process succeeds even if the notification fails. We make a
"good faith" effort to send a notification, but the purchase process will
succeed regardless of the outcome of sending a notification.