Feature: Add admin db class for adding/updating locations
Four location functions:
1. Adding
2. Updating
3. Disabling Payments
4. Reenabling Payments

This stores each location in the db by gps coordinates. First the location is
added to the vendor node and that id is used to write to each of the gps
coordinate nodes.

Each time a location is added the `lastModified` is updated with the current
unix time. This allows the client to check the lastModified time to determine
if a node needs to be downloaded again.

Updating a vendor will be a challenge if there are many vendors as queries in
firebase can be slow, but we can revisit this in the future.