Admin: Add location input utilities
Enables adding/updating/disabling/reenabling a location with the firebase
admin db.

Location name and address are queried against the google maps api for
gps coordinates and a formatted address. Google maps also has the ability to
differentiate gps coordinates for businesses with the same street address.

This is designed to be executed with `ts-node`.