Feature: Pull locations near/closest to user
Request device gps coordinates and use them to pull down locations nearest the
user. Uses the nodes formatted in `addLocation`.

Also adds a function to query the narrowest node to find the location where
the user is located. Used for verifying redemption.

Right now this only queries one node. In the feature we could query a square
of nodes around the user in case the user is in the corner of one node, but
note this is O(x^3) in the best case.