Refactor: Improve reliability of firebase listeners
DRY up listening for server events, wrapping them into
`listenForChangeOnUrl`.

Add `transactionFailed` and `transactionComplete` functions which parse
the latest transaction status for a failed or complete status.

Add `listenUntilTransactionSuccessOrFailure` which loops the transaction
listener until the status shows failure or completion.

This should be much more robust than the previous implementation which
listened for an exact number of events.