Feature: Pass redeem transaction status from server to client
Oddly enough this happens too fast on the server. By the time the client
starts listening the second status update is complete. To fix this the
server waits in a `setTimeout` for a small amount of time, then updates
the `redeemTransactionStatus` node.

There were also problems with the `TransactionStatus` parsing plugins.
`transactionFinished` returns if the `TransactionStatus` indicates a
failure or a completion, indicating the client should stop listening.