Feature: Use viewport AND square footage to determine user location
Google places `viewport` is not the actual location viewport but a map
viewport, thus it is not accurate for determining an exact location.

This uses the viewport as a filter, then calculates the distance
between the user and the business location. The business square
footage is converted a radius to compare against the distance result.