Android: Use different facebook api key for dev and production
For production builds:
`./gradlew assembleRelease -Pbuild_key_type=production`

For dev builds the default is dev so no special project input is needed.

The actual key is stored in `.gradle/gradle.properties`.

Using the key is the manifest was difficult due to android not having
types for manifest variables. To make android interpret this number as a
string we append a `\0` null character to the end of the variable. This
is undocumented, but works for now.