Feature: Add production build script.
This script performs the necessary tasks to build a production ready
version of the android app apk (ios still needs work for automation) and
update the server.

Tasks are performed in order, and the script is aborted if any one
returns a failing status code.

From a high level:

1. Gets the version number from `package.json`.
2. Tells the user to turn on the android emulator with gps on.
3. Verifies the correct files exist/don't exist in `BuzzPath`, returning
an object with valid directories and filenames.
4. Sets up `buzzBuilder` object for performing app specific build tasks.
5. Runs javascript tests with `npm test`.
6. Runs Android tests with `./gradlew connectedAndroidTest`.
7. Builds the android dev apk.
8. Builds the android production apk. This swaps the
`google-services.json` and `google-services-production.json` files to
use production firebase services. The gradle command is passed
`-Pbuild_key_type=production` to use the production facebook app id.
9. Unzips each android bundle, and verifies no envelope keys are present
in the js bundle.
10. Updates the firebase rules for the dev and production apps
11. Writes the env variables to each heroku instance. Env keys prefixed
with `PRODUCTION_` are passed to the production queue server.
12. Git push dev and production heroku deploys.