Feature: Add reducer for `banners`
A "Banner" is a non intrusive message displayed below the branding
header. Currently it is being used to display a "No Internet Connection"
message when the user tries to refresh `ScrollView`s that require an
internet connection. I see the banner as a cross platform alternative to
Android's `Toast` functionality.