Compiler: Check for unused local variables and imports
This is the first step in cleaning up the codebase using the typescript
compiler.

`tsc` configured with `strict: true` catches way too many errors right
now, but `noUnusedLocals` and `noImplicitThis` are easy to fix and
seem like a good starting point.

In the future I plan to fix all errors with `strict: true` enabled.