Feature: Add step 1 of 4 step redeem process, Location selection
This refactors, cleans up, and removes much of the previous redeem logic
and adds a location picker to the first view of the redeem process.

This redeem location picker puts the burden on the user to select their
location correctly, rather than relying on the users gps coordinates. A
nice feature here is the "Other" line allowing the user to see a message
explaining why their location might not be in the list.