Feature: Add step 4 of 4 step redeem process: Redeem Completion view
This is a simple view showing a summary of what just happened. It is
important to show enough information to prevent screenshots, but not be
so complicated that the information cannot be parsed at a glance.

Before the view can close the user must confirm they received their
drinks before the view closes.

The time changes from green to red based on the distance between the
redeem time and the actual time.