Refactor: Migrate to `FlatList` in `History` view.
`FlatList` is supposed to perform better than `ListView`. The big catch
with `FlatList` is it needs a pure component for the render function.
Essentially a pure component is self contained, not calling on variables
outside its scope (except for imported variables?). I don't have any
numbers but the render feels faster.

This breaks out a `BevFlatList` class for future refactoring of
components using `ListView`.