Fix: Failure on logging out
When the `MainTabView` component unmounts the `turnOffNotifications`
function is called. This fails hard, and cryptically, when the component
unmounts after the user logs out.

This checks if the user is logged in before attempting to turn off
notifications. This also removes the `stopNotificationListener` from
`MainTabView` because we should only explicitly stop notifications,
either when the user logs out of the user toggles the notification
setting.