Fix: Improve reliability of purchase and redeem server calls
This uses starts a firebase database listener on a status node that is
updated by the server. When a status node is changed on the server the
app dispatches that change to the state.  We dispatch this status to the
appropriate location in the state and check if we need to stop the
listener with the `transactionFinished` function. Upon completion, or
exception we stop the listener.

This is more reliable because there is no potential for a listener being
setup after a change has been made on the server. The previous
implementation opened and closed a listener between each status update.
Now we just listen until the transaction is finished, which is faster,
simpler and more reliable.